refactor: normalize Inertia render paths and remove orphan pages
- Consolidate Accessibility.vue + accessibility/Index.vue into a single
  self-contained Inertia page with AppLayout, Head, and breadcrumbs
- Update AccessibilityController to render 'Accessibility/Index'
- Fix double-slash import in reports/sales/Index.vue
- Remove stale commented import in Menus/Index.vue
- Delete 16 confirmed dead files: stale scaffolds (Welcome, TerminalSession,
  Branches), duplicate roles pages, entire old pages/branch/ and pages/menu/
  implementations (superseded by components/Menus/ and branches/IndexBranches)
